refactor: extract shared and environment stack entry points
Split monolithic Program.cs into a thin dispatcher that branches on
environment, delegating to SharedStack.cs (for the shared Pulumi stack)
or EnvironmentStack.cs (for dev/prod stacks). No infrastructure changes.
