+++ infra/SharedStack.cs
@@ -1,0 +1,9 @@
+using Pulumi;
+
+public static class SharedStack
+{
+    public static Dictionary<string, object?> Run(Config config, InputMap<string> tags)
+    {
+        return new Dictionary<string, object?>();
+    }
+}
